C2LC-454: Move background information object in BackgroundInfo.js to worlds object in Worlds.js
