+++ src/Scene.js
@@ -4,8 +4,7 @@ import React from 'react';
 import CharacterState from './CharacterState';
 import Character from './Character';
 import SceneDimensions from './SceneDimensions';
-import { getWorldProperties } from './Worlds';
-import { getBackgroundInfo } from './BackgroundInfo';
+import { getBackgroundInfo, getWorldProperties } from './Worlds';
 import { injectIntl } from 'react-intl';
 import type {IntlShape} from 'react-intl';
 import './Scene.scss';
